feat: BottomNavigationView 추가 and Fragment 생성
