+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PetCalendar">
-        <activity android:name="com.minuk.petcalendar.MainActivity">
+        <activity android:name="com.minuk.petcalendar.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
