+++ app/src/main/java/com/minuk/petcalendar/calendar/CalendarFragment.kt
@@ -1,0 +1,25 @@
+package com.minuk.petcalendar.calendar
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.minuk.petcalendar.R
+
+class CalendarFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_calendar, container, false)
+    }
+
+    companion object {
+
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) = CalendarFragment()
+    }
+}
