+++ app/src/main/res/drawable/bg_selection_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="@color/main_accent" />
+    <item android:color="@color/main_surface"  />
+</selector>
