+++ app/src/main/res/drawable/ic_calendar.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M16.4109,2.7686L16.4119,3.5182C19.1665,3.7341 20.9862,5.6112 20.9891,8.4897L21,16.9155C21.0039,20.054 19.0322,21.985 15.8718,21.99L8.1519,22C5.0112,22.004 3.0148,20.027 3.0109,16.8795L3,8.5527C2.9961,5.6552 4.7515,3.7831 7.5062,3.5302L7.5052,2.7806C7.5042,2.3408 7.83,2.01 8.2644,2.01C8.6989,2.009 9.0247,2.3388 9.0257,2.7786L9.0267,3.4783L14.8914,3.4703L14.8904,2.7706C14.8894,2.3308 15.2152,2.001 15.6497,2C16.0742,1.999 16.4099,2.3288 16.4109,2.7686ZM4.5215,8.8616L19.4696,8.8416V8.4917C19.4272,6.3428 18.349,5.2154 16.4138,5.0475L16.4148,5.8171C16.4148,6.2469 16.0801,6.5877 15.6556,6.5877C15.2212,6.5887 14.8943,6.2489 14.8943,5.8191L14.8934,5.0095L9.0286,5.0175L9.0296,5.8261C9.0296,6.2569 8.7048,6.5967 8.2704,6.5967C7.8359,6.5977 7.5091,6.2589 7.5091,5.8281L7.5082,5.0585C5.5829,5.2514 4.5175,6.3828 4.5205,8.5507L4.5215,8.8616ZM15.2399,13.4043V13.4153C15.2498,13.8751 15.625,14.2239 16.0801,14.2139C16.5244,14.2029 16.8789,13.8221 16.869,13.3623C16.8483,12.9225 16.4918,12.5637 16.0485,12.5647C15.5944,12.5747 15.2389,12.9445 15.2399,13.4043ZM16.0554,17.892C15.6013,17.882 15.235,17.5032 15.234,17.0435C15.2241,16.5837 15.5884,16.2029 16.0426,16.1919H16.0525C16.5165,16.1919 16.8927,16.5707 16.8927,17.0405C16.8937,17.5102 16.5185,17.891 16.0554,17.892ZM11.1721,13.4203C11.1919,13.8801 11.568,14.2389 12.0222,14.2189C12.4665,14.1979 12.821,13.8181 12.8012,13.3583C12.7903,12.9085 12.425,12.5587 11.9807,12.5597C11.5266,12.5797 11.1711,12.9605 11.1721,13.4203ZM12.0262,17.8471C11.572,17.8671 11.1968,17.5082 11.1761,17.0485C11.1761,16.5887 11.5305,16.2089 11.9847,16.1879C12.429,16.1869 12.7953,16.5367 12.8052,16.9855C12.8259,17.4463 12.4705,17.8261 12.0262,17.8471ZM7.1043,13.4553C7.1241,13.915 7.5002,14.2749 7.9544,14.2539C8.3987,14.2339 8.7532,13.8531 8.7324,13.3933C8.7226,12.9435 8.3573,12.5937 7.912,12.5947C7.4578,12.6147 7.1033,12.9955 7.1043,13.4553ZM7.9584,17.8521C7.5042,17.8731 7.129,17.5132 7.1083,17.0535C7.1073,16.5937 7.4627,16.2129 7.9169,16.1929C8.3612,16.1919 8.7275,16.5417 8.7374,16.9915C8.7581,17.4513 8.4037,17.8321 7.9584,17.8521Z"
+      android:fillColor="#1D6A96"
+      android:fillType="evenOdd"/>
+</vector>
