+++ app/src/main/res/drawable/ic_setting.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M20.4023,13.5801C20.76,13.7701 21.036,14.0701 21.2301,14.3701C21.6083,14.9901 21.5776,15.7501 21.2097,16.4201L20.4943,17.6201C20.1162,18.2601 19.411,18.6601 18.6855,18.6601C18.3278,18.6601 17.9292,18.5601 17.6022,18.3601C17.3365,18.1901 17.0299,18.1301 16.7029,18.1301C15.6911,18.1301 14.8429,18.9601 14.8122,19.9501C14.8122,21.1001 13.872,22.0001 12.6968,22.0001H11.3069C10.1215,22.0001 9.1812,21.1001 9.1812,19.9501C9.1608,18.9601 8.3126,18.1301 7.3008,18.1301C6.9636,18.1301 6.657,18.1901 6.4015,18.3601C6.0745,18.5601 5.6657,18.6601 5.3183,18.6601C4.5824,18.6601 3.8773,18.2601 3.4992,17.6201L2.794,16.4201C2.4159,15.7701 2.3955,14.9901 2.7736,14.3701C2.9371,14.0701 3.2437,13.7701 3.5912,13.5801C3.8773,13.4401 4.0613,13.2101 4.235,12.9401C4.746,12.0801 4.4394,10.9501 3.5707,10.4401C2.559,9.8701 2.2319,8.6001 2.8145,7.6101L3.4992,6.4301C4.0919,5.4401 5.3591,5.0901 6.3811,5.6701C7.2702,6.1501 8.425,5.8301 8.9462,4.9801C9.1097,4.7001 9.2017,4.4001 9.1812,4.1001C9.1608,3.7101 9.2732,3.3401 9.4674,3.0401C9.8455,2.4201 10.5302,2.0201 11.2763,2.0001H12.7172C13.4735,2.0001 14.1582,2.4201 14.5363,3.0401C14.7203,3.3401 14.8429,3.7101 14.8122,4.1001C14.7918,4.4001 14.8838,4.7001 15.0473,4.9801C15.5685,5.8301 16.7233,6.1501 17.6226,5.6701C18.6344,5.0901 19.9118,5.4401 20.4943,6.4301L21.179,7.6101C21.7718,8.6001 21.4447,9.8701 20.4228,10.4401C19.5541,10.9501 19.2475,12.0801 19.7687,12.9401C19.9322,13.2101 20.1162,13.4401 20.4023,13.5801ZM9.1097,12.0101C9.1097,13.5801 10.4076,14.8301 12.0121,14.8301C13.6165,14.8301 14.8838,13.5801 14.8838,12.0101C14.8838,10.4401 13.6165,9.1801 12.0121,9.1801C10.4076,9.1801 9.1097,10.4401 9.1097,12.0101Z"
+      android:fillColor="#1D6A96"
+      android:fillType="evenOdd"/>
+</vector>
