+++ app/src/main/res/menu/navigation_item.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/action_calendar"
+        android:title="@string/menu_calendar"
+        android:icon="@drawable/ic_calendar" />
+
+    <item android:id="@+id/action_document"
+        android:title="@string/menu_document"
+        android:icon="@drawable/ic_document" />
+
+    <item android:id="@+id/action_setting"
+        android:title="@string/menu_setting"
+        android:icon="@drawable/ic_setting" />
+</menu>
