+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="main_calendarDate_textSize">18dp</dimen>
+</resources>
